+++ SharpNoise/Modules/Simplex.cs
@@ -5,8 +5,6 @@
 /// </summary>
 public class Simplex : Module
 {
-    public override ReadOnlySpan<Module> SourceModules => [];
-
     private readonly struct Grad(double x, double y, double z)
     {
         public readonly double X = x, Y = y, Z = z;