+++ SharpNoise/NoiseMath.cs
@@ -141,28 +141,6 @@ public static class NoiseMath
     }
 
     /// <summary>
-    /// Maps a value onto a quintic S-curve.
-    /// </summary>
-    /// <param name="a">The value to map onto a quintic S-curve.</param>
-    /// <returns>The mapped value.</returns>
-    /// <remarks>
-    /// The first derivitive of a quintic S-curve
-    /// is zero at <paramref name="a"/> = 0.0 and
-    /// <paramref name="a"/> = 1.0
-    ///
-    /// The second derivitive of a quintic S-curve
-    /// is zero at <paramref name="a"/> = 0.0 and
-    /// <paramref name="a"/> = 1.0
-    /// </remarks>
-    public static double SCurve5(double a)
-    {
-        var a3 = a * a * a;
-        var a4 = a3 * a;
-        var a5 = a4 * a;
-        return (6.0 * a5) - (15.0 * a4) + (10.0 * a3);
-    }
-
-    /// <summary>
     /// Converts latitude/longitude coordinates on a unit sphere into 3D
     /// Cartesian coordinates.
     /// </summary>
@@ -181,48 +159,6 @@ public static class NoiseMath
         (z, x) = Math.SinCos(DegToRad * lon);
         x *= r;
         z *= r;
-    }
-
-    /// <summary>
-    /// Clamps a value onto a clamping range.
-    /// </summary>
-    /// <param name="value">The value to clamp.</param>
-    /// <param name="lowerBound">The lower bound of the clamping range.</param>
-    /// <param name="upperBound">The upper bound of the clamping range.</param>
-    /// <returns>
-    /// - <paramref name="value" /> if <paramref name="value" /> lies between <paramref name="lowerBound" /> and <paramref name="upperBound" />.
-    /// - <paramref name="lowerBound" /> if <paramref name="value" /> is less than <paramref name="lowerBound" />.
-    /// - <paramref name="upperBound" /> if <paramref name="value" /> is greater than <paramref name="upperBound" />.
-    /// </returns>
-    public static int Clamp(int value, int lowerBound, int upperBound)
-    {
-        return value > upperBound ? upperBound : (value < lowerBound ? lowerBound : value);
-    }
-
-    /// <summary>
-    /// Clamps a value onto a clamping range.
-    /// </summary>
-    /// <param name="value">The value to clamp.</param>
-    /// <param name="lowerBound">The lower bound of the clamping range.</param>
-    /// <param name="upperBound">The upper bound of the clamping range.</param>
-    /// <returns>
-    /// - <paramref name="value" /> if <paramref name="value" /> lies between <paramref name="lowerBound" /> and <paramref name="upperBound" />.
-    /// - <paramref name="lowerBound" /> if <paramref name="value" /> is less than <paramref name="lowerBound" />.
-    /// - <paramref name="upperBound" /> if <paramref name="value" /> is greater than <paramref name="upperBound" />.
-    /// </returns>
-    public static double Clamp(double value, double lowerBound, double upperBound)
-    {
-        return value > upperBound ? upperBound : (value < lowerBound ? lowerBound : value);
-    }
-
-    /// <summary>
-    /// Swaps two values.
-    /// </summary>
-    /// <param name="a">A variable containing the first value.</param>
-    /// <param name="b">A variable containing the second value.</param>
-    public static void Swap<T>(ref T a, ref T b)
-    {
-        (b, a) = (a, b);
     }
 
     /// <summary>